feat: add route to update a salad
