+++ apps/api/src/main.ts
@@ -3,8 +3,7 @@ import cors from 'cors';
 import config from './config';
 import { spellcheckWord } from './spellchecker';
 import { query } from './db';
-import { saladGenerator } from './salad-calculator';
-import { wordGenerator } from './word-generator';
+import { generateWordSalad } from './salad-calculator';
 
 const host = process.env.HOST ?? 'localhost';
 const port = process.env.PORT ? Number(process.env.PORT) : 3000;
@@ -51,36 +50,28 @@ app.post('/spellcheck', async (req, res) => {
 });
 
 app.post('/generate-salad', async (_req, res) => {
-  const { rows } = await query({
-    text: 'select initial_word from salads',
-  });
-  // query existing words
-  const usedWords = rows?.map((row) => row.initial_word);
-  // generate unused initial word
-  const initialWord = wordGenerator(usedWords);
-
-  let solutionString: string | undefined = undefined;
-
-  while (!solutionString) {
-    // generate word salads from new word
-    const wordSalads = saladGenerator(initialWord);
-    if (wordSalads.length > 0) {
-      solutionString = wordSalads;
-    }
-  }
-
-  // catch empty solution set
-  if (!solutionString) {
-    throw new Error('Failed to generate a solution set');
-  }
-
+  const { initialWord, solutionSet } = await generateWordSalad();
   // insert into postgres
   const current_date = new Date();
   const yyyyMmDd = current_date.toISOString().split('T')[0];
   const formatted = `${yyyyMmDd} 00:00:00`;
   await query({
     text: 'insert into salads (initial_word, par, date, solution_set) values ($1, 6, $2, $3)',
-    params: [initialWord, formatted, solutionString],
+    params: [initialWord, formatted, solutionSet],
+  });
+  res.send({ initialWord });
+});
+
+app.post('/update-salad', async (_req, res) => {
+  const { initialWord, solutionSet } = await generateWordSalad();
+  const { rows } = await query({
+    text: 'select id from salads order by salad_number desc limit 1',
+  });
+  const id = rows[0]?.id;
+  // update postgres
+  await query({
+    text: 'update salads set initial_word = $1, solution_set = $2 where id = $3',
+    params: [initialWord, solutionSet, id],
   });
   res.send({ initialWord });
 });
